Added fade in, made navbar non-iframe element, updated cv
Added a fade in (with a 5 second variant and a .5 second variant) that can be assigned to elements using class tags.
Made nav bar a non iframe object to reduce  annoying pop in problem that didnt seem to go away with waiting till the page was loaded and making them visable.
Updated cv file with my most up to date copy.